reorganized code into multiple crates to avoid large builds for core and py.
