+++ Cargo.toml
@@ -1,19 +1,14 @@
 [workspace]
+members = ["core", "cli", "py"]
 
-[package]
-name = "rtpa_core"
-version = "0.0.1"
-edition = "2021"
+# Optional: Set the default member for cargo commands
+default-members = ["cli"]
 
-[lib]
-name = "rtpa_core"
-crate-type = ["cdylib", "rlib"]
-
-
-[dependencies]
+# Optional: Shared dependencies to ensure version consistency
+[workspace.dependencies]
 arrow = { version = "54.3.1", features = ["ipc", "pyarrow"] }
 axum = "0.7.7"
-bytes = "1.7.1"
+bytes = "1.7.1"                                                     # This was missing
 chrono = { version = "0.4.40", features = ["serde"] }
 clap = { version = "4.0", features = ["derive"] }
 serde = { version = "1.0", features = ["derive"] }
@@ -26,28 +21,3 @@ tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 tracing-appender = "0.2"
 rayon = "1.10.0"
 pyo3 = { version = "0.23.3", features = ["extension-module"] }
-
-
-[dev-dependencies]
-criterion = { version = "0.5.1", features = ["html_reports"] }
-reqwest = { version = "0.12.8", features = ["json"] }
-tokio = { version = "1.0", features = ["full"] }
-chrono = { version = "0.4.40", features = ["serde"] }
-arrow = { version = "54.3.1", features = ["ipc"] }
-num_cpus = "1.16.0"
-random = "0.14.0"
-rand = "0.9.0"
-
-
-[[bench]]
-name = "arrow_latency"
-harness = false
-
-[[bench]]
-name = "accumulator_throughput"
-harness = false
-
-
-[[test]]
-name = "integration_test"
-path = "tests/pdc_buffer_test.rs"
