+++ build.md
@@ -1,0 +1,8 @@
+# RTPA-core
+
+Build instructions for RTPA-core
+
+To build all crates, run `cargo build` as normal.
+To build individual crates, run `cargo build -p <crate_name>`.
+
+### Core
