+++ cli/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "rtpa_cli"
+version = "0.0.1"
+edition = "2021"
+
+[[bin]]
+name = "rtpa"
+path = "src/main.rs"
+
+[dependencies]
+# Reference the core library
+rtpa_core = { path = "../core" }
+
+# CLI-specific dependencies
+axum = { workspace = true }
+arrow = { workspace = true }
+bytes = { workspace = true }
+chrono = { workspace = true }
+clap = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+tokio = { workspace = true }
+tower = { workspace = true }
+tower-http = { workspace = true }
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true }
+tracing-appender = { workspace = true }
