+++ cli/src/arrow_utils.rs
