+++ cli/src/lib.rs
@@ -6,17 +6,3 @@ pub mod frames;
 pub mod pdc_buffer_server;
 pub mod pdc_client;
 pub mod pdc_server;
-
-// Declare the accumulators module
-pub mod accumulator {
-    pub mod manager;
-
-    pub mod sparse;
-}
-
-pub mod ieee_c37_118;
-pub mod utils;
-
-pub mod pdc_buffer;
-
-pub mod pdc_buffer_py;
