+++ core/src/accumulator/manager.rs
