+++ core/src/accumulator/sparse.rs
