+++ core/src/ieee_c37_118/common.rs
