+++ core/src/lib.rs
@@ -1,0 +1,11 @@
+// Declare the accumulators module
+pub mod accumulator {
+    pub mod manager;
+
+    pub mod sparse;
+}
+
+pub mod ieee_c37_118;
+pub mod utils;
+
+pub mod pdc_buffer;
